requirements.txt
segno
pillow